+++ src/t1/i2cimpl.rs
@@ -17,11 +17,27 @@ mod nrf52832 {
     }
 }
 
-#[cfg(feature = "lpc55")]
-mod lpc55 {
+#[cfg(feature = "lpc55-v0.3")]
+mod lpc55_03 {
     use crate::t1::I2CErrorNack;
 
     use lpc55_hal::drivers::i2c::Error;
+
+    impl I2CErrorNack for Error {
+        fn is_address_nack(&self) -> bool {
+            matches!(self, Error::NackAddress)
+        }
+        fn is_data_nack(&self) -> bool {
+            matches!(self, Error::NackData)
+        }
+    }
+}
+
+#[cfg(feature = "lpc55-v0.4")]
+mod lpc55_04 {
+    use crate::t1::I2CErrorNack;
+
+    use lpc55_hal_04::drivers::i2c::Error;
 
     impl I2CErrorNack for Error {
         fn is_address_nack(&self) -> bool {