+++ CHANGELOG.md
@@ -5,6 +5,10 @@ SPDX-License-Identifier: CC0-1.0
 
 # Changelog
 
+## Unreleased
+
+- Add support for `lpc55-hal` 0.4 behind a `lpc55-v0.4` feature
+
 ## [v0.1.6][] (2025-08-13)
 
 - Fix `ExportObject` for large keys (NIST-P521 and brainpoolp512) ([#16][])